Move test to correct class
